# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Vitas Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get all banners",
                "description": "Get all banners",
                "operationId": "b74baafbdca35c012150caf9d9101f3a",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Create a new banner",
                "description": "Create a new banner",
                "operationId": "8cb8a9c999b6c10a6ba2743aaa8c896e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get banner by id",
                "description": "Get banner by id",
                "operationId": "47de669e7753e8c387205f41f0ae7e4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Banners"
                ],
                "summary": "Update a new banner",
                "description": "Update a new banner",
                "operationId": "21c37b31a3303f82171c89f31f694312",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete banner by id",
                "description": "Delete banner by id",
                "operationId": "9d7ceb33a50c1b8cad1292d4dde3c3b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings",
                "description": "Get all bookings",
                "operationId": "4b6323b45b01fd18933b9480fafef1e7",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "description": "Create a new booking",
                "operationId": "b8d9d18a7b29e7008123a0813237d5b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}